web-api/face_detection_age_classification/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib
colorama==0.4.6
confection==0.1.4
configobj==5.0.8
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dulwich==0.21.7
everett==3.1.0
extensions==0.4
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2023.12.2
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
preshed==3.0.9
protobuf==4.25.1
psutil==5.9.7
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic
pydantic_core
Pygments==2.17.2
pyparsing==3.1.1
python-box==6.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.13.1
semantic-version==2.10.0
sentry-sdk==1.39.1
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
srsly==2.4.8
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-intel
termcolor==2.4.0
thinc==8.2.2
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typer==0.9.0
typing==3.7.4.3
typing_extensions
tzdata==2023.4
ultralytics==8.0.237
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
websocket-client==1.3.3
Werkzeug==3.0.1
wrapt==1.16.0
wurlitzer==3.0.3